+++ inst/examples/ex-bootstrap.R
@@ -7,5 +7,5 @@ bootstrap(x, do = mean, n = 100)
 quant <- function(x) { quantile(x, probs = c(0.25, 0.75)) }
 bootstrap(x, n = 100, do = mean, f = quant)
 
-## Get the n bootstrap values
+## Get the n bootstrap estimates
 bootstrap(x, n = 100, do = mean, f = function(x) { x })